Moved waitForStatus(withCodes:) to StageGroup
